trace: cheri: fix syscall_exit tracepoints for cheri
Generally, a tracepoint is associated with one or more probe functions.
When the tracepoint is triggered, all probe functions are called.

For tracepoints that map to an event, the kernel installs a probe
function that logs into the ringbuffer. It's possbile to install
further probe functions, e.g. by loading modules.

On a cheri system, the tracepoints and probe functions should use
capabilities. Only the logging code should reduce them to addresses.

Update the generated syscall_exit tracepoints accordingly. For cheri, a
sycall may return an integer value or a capability.

trace_sys_exit triggers the tracepoint, its parameters are defined by
TP_PROTO. ftrace_syscall_exit is the default probe function that writes
a log message. All of them should use capabilities, ftrace_syscall_exit
should reduce the capabilities to addresses.

struct syscall_trace_exit defines a ringbuffer entry. We need addresses
here.

+++ include/linux/entry-common.h
@@ -66,7 +66,7 @@ static __always_inline int arch_ptrace_report_syscall_entry(struct pt_regs *regs
 
 bool syscall_user_dispatch(struct pt_regs *regs);
 long trace_syscall_enter(struct pt_regs *regs, long syscall);
-void trace_syscall_exit(struct pt_regs *regs, long ret);
+void trace_syscall_exit(struct pt_regs *regs, intptr_t ret);
 
 static inline void syscall_enter_audit(struct pt_regs *regs, long syscall)
 {
